Break into VaultElement
