+++ BTCPayServer/BTCPayServer.csproj
@@ -54,7 +54,7 @@
     <PackageReference Include="NBitcoin" Version="8.0.8" />
     <PackageReference Include="YamlDotNet" Version="8.0.0" />
     <PackageReference Include="BIP78.Sender" Version="0.2.4" />
-    <PackageReference Include="BTCPayServer.Hwi" Version="2.0.2" />
+    <PackageReference Include="BTCPayServer.Hwi" Version="2.0.6" />
     <PackageReference Include="BTCPayServer.Lightning.All" Version="1.6.9" />
     <PackageReference Include="CsvHelper" Version="32.0.3" />
     <PackageReference Include="Dapper" Version="2.1.35" />
