+++ BTCPayServer/Blazor/VaultBridge/Elements/FeedbackType.cs
@@ -1,0 +1,8 @@
+﻿namespace BTCPayServer.Blazor.VaultBridge.Elements;
+
+public enum FeedbackType
+{
+    Loading,
+    Success,
+    Failed
+}
