+++ BTCPayServer/Blazor/VaultBridge/IController.cs
@@ -1,0 +1,9 @@
+﻿using System.Threading;
+using System.Threading.Tasks;
+
+namespace BTCPayServer.Blazor.VaultBridge;
+public interface IController
+{
+    Task Run(VaultBridgeUI ui, CancellationToken cancellationToken);
+}
+
