+++ BTCPayServer/Blazor/VaultBridge/VaultElement.cs
@@ -1,0 +1,11 @@
+﻿using Microsoft.AspNetCore.Components;
+using Microsoft.AspNetCore.Components.Rendering;
+
+namespace BTCPayServer.Blazor.VaultBridge;
+
+public class VaultElement
+{
+    protected virtual void BuildRenderTree(RenderTreeBuilder builder) { }
+
+    public RenderFragment RenderFragment => BuildRenderTree;
+}
