--- BTCPayServer/Views/Shared/LocalhostBrowserSupport.cshtml
@@ -1,22 +1,0 @@
-<div id="walletAlert" class="alert alert-warning alert-dismissible my-4" style="display:none;" role="alert">
-    <button type="button" class="btn-close" data-bs-dismiss="alert" aria-label="@StringLocalizer["Close"]">
-        <vc:icon symbol="close" />
-    </button>
-    <span id="alertMessage"></span>
-</div>
-<script>
-		var alertMsg = document.getElementById("alertMessage");
-		var walletAlert = document.getElementById("walletAlert");
-		var isSafari = window.safari !== undefined;
-		if (isSafari)
-		{
-			alertMsg.innerHTML = "Safari doesn't support BTCPay Server Vault. Please use a different browser. (<a class=\"alert-link\" href=\"https://bugs.webkit.org/show_bug.cgi?id=171934\" target=\"_blank\" rel=\"noreferrer noopener\">More information</a>)";
-			walletAlert.style.display = null;
-		}
-		var isBrave = navigator.brave !== undefined;
-		if (isBrave)
-		{
-			alertMsg.innerHTML = "Brave supports BTCPay Server Vault, but you need to disable Brave Shields. (<a class=\"alert-link\" href=\"https://www.updateland.com/how-to-turn-off-brave-shields/\" target=\"_blank\" rel=\"noreferrer noopener\">More information</a>)";
-			walletAlert.style.display = null;
-		}
-</script>
