--- BTCPayServer/Views/Shared/VaultElements.cshtml
@@ -1,69 +1,0 @@
-<template id="VaultConnection">
-    <div class="vault-feedback vault-feedback1 mb-2 d-flex align-items-center">
-        <vc:icon symbol="dots" css-class="vault-feedback-icon d-none"/>
-        <span class="vault-feedback-content flex-grow"></span>
-    </div>
-    <div class="vault-feedback vault-feedback2 mb-2 d-flex align-items-center">
-        <vc:icon symbol="dots" css-class="vault-feedback-icon d-none"/>
-        <span class="vault-feedback-content flex-grow"></span>
-    </div>
-    <div class="vault-feedback vault-feedback3 mb-2 d-flex align-items-center">
-        <vc:icon symbol="dots" css-class="vault-feedback-icon d-none"/>
-        <span class="vault-feedback-content flex-grow"></span>
-    </div>
-    <div class="vault-feedback vault-feedback4 mb-2 d-flex align-items-center">
-        <vc:icon symbol="dots" css-class="vault-feedback-icon d-none"/>
-        <span class="vault-feedback-content flex-grow"></span>
-    </div>
-	<div class="vault-feedback vault-feedback5 mb-2 d-flex align-items-center">
-        <vc:icon symbol="dots" css-class="vault-feedback-icon d-none"/>
-        <span class="vault-feedback-content flex-grow"></span>
-	</div>
-    <div id="pin-input" class="mt-4" style="display: none;">
-        <div class="row">
-            <div class="col">
-                <div class="input-group mb-2">
-                    <input id="pin-display" type="text" class="form-control" readonly>
-                    <div id="pin-display-delete" class="input-group-text cursor-pointer">
-                        <vc:icon symbol="actions-remove" />
-                    </div>
-                </div>
-            </div>
-        </div>
-        <div class="row">
-            <div class="col"><div class="pin-button" id="pin-7"></div></div>
-            <div class="col"><div class="pin-button" id="pin-8"></div></div>
-            <div class="col"><div class="pin-button" id="pin-9"></div></div>
-        </div>
-        <div class="row">
-            <div class="col"><div class="pin-button" id="pin-4"></div></div>
-            <div class="col"><div class="pin-button" id="pin-5"></div></div>
-            <div class="col"><div class="pin-button" id="pin-6"></div></div>
-        </div>
-        <div class="row">
-            <div class="col"><div class="pin-button" id="pin-1"></div></div>
-            <div class="col"><div class="pin-button" id="pin-2"></div></div>
-            <div class="col"><div class="pin-button" id="pin-3"></div></div>
-        </div>
-    </div>
-    <div id="passphrase-input" class="mt-4" style="display: none;">
-        <div class="form-group">
-            <label for="Password" class="form-label" text-translate="true">Passphrase (Leave empty if there isn't any passphrase)</label>
-            <div class="input-group">
-                <input id="Password" type="password" class="form-control">
-                <button type="button" class="btn btn-secondary px-3 only-for-js" title="@StringLocalizer["Toggle passphrase visibility"]" data-toggle-password="#Password">
-                    <vc:icon symbol="actions-show" />
-                </button>
-            </div>
-        </div>
-        <div class="form-group">
-            <label for="PasswordConfirmation" class="form-label" text-translate="true">Passphrase confirmation</label>
-            <div class="input-group">
-                <input id="PasswordConfirmation" type="password" class="form-control">
-                <button type="button" class="btn btn-secondary px-3 only-for-js" title="@StringLocalizer["Toggle passphrase visibility"]" data-toggle-password="#PasswordConfirmation">
-                    <vc:icon symbol="actions-show" />
-                </button>
-            </div>
-        </div>
-    </div>
-</template>
