+++ BTCPayServer/Views/UIPullPayment/SetupBoltcard.cshtml
@@ -1,7 +1,7 @@
 @using BTCPayServer.Abstractions.TagHelpers
 @using BTCPayServer.Blazor
+@using BTCPayServer.Blazor.VaultBridge
 @using Microsoft.AspNetCore.Mvc.TagHelpers
-
 @using BTCPayServer.TagHelpers
 @model SetupBoltcardViewModel
 @{
@@ -30,8 +30,8 @@
 
 <div class="row mt-5 mb-4">
     <div class="col-md-8 mx-auto">
-        <component type="typeof(BTCPayServer.Blazor.VaultBridgeUI)"
-                   param-Controller="@(new VaultBridgeUI.NFCController()
+        <component type="typeof(VaultBridgeUI)"
+                   param-Controller="@(new NFCController()
                                      {
                                          NewCard = Model.NewCard,
                                          BoltcardUrl = Model.BoltcardUrl,
