+++ BTCPayServer/Views/UIWallets/WalletSendVault.cshtml
@@ -1,6 +1,6 @@
 @using BTCPayServer.Abstractions.TagHelpers
-@using BTCPayServer.Blazor
-@using BTCPayServer.Controllers
+@using BTCPayServer.Blazor.VaultBridge
+@using BTCPayServer.TagHelpers
 @using Microsoft.AspNetCore.Mvc.TagHelpers
 @model WalletSendVaultModel
 @{
@@ -22,8 +22,8 @@
 
 <div class="row mt-5 mb-4">
     <div class="col-md-8 mx-auto">
-        <component type="typeof(BTCPayServer.Blazor.VaultBridgeUI)"
-                   param-Controller="@(new VaultBridgeUI.SignHWIController()
+        <component type="typeof(VaultBridgeUI)"
+                   param-Controller="@(new SignHWIController()
                                       {
                                           StoreId = WalletId.Parse(walletId).StoreId,
                                           CryptoCode = WalletId.Parse(walletId).CryptoCode,
